split import CSV in 2 process => Verifier and Treatment && implement Typo and Duplicate
