+++ src/BeneficiaryBundle/Controller/HouseholdController.php
@@ -132,37 +132,46 @@ class HouseholdController extends Controller
         if (!$request->query->has('step'))
             return new Response('You must specify the current level.');
         $step = $request->query->get('step');
+        if ($request->query->has('token'))
+            $token = $request->query->get('token');
+        else
+            $token = null;
+
+        $contentJson = $request->request->all();
+        $countryIso3 = $contentJson['__country'];
+        unset($contentJson['__country']);
+        /** @var HouseholdCSVService $householdService */
+        $householdService = $this->get('beneficiary.household_csv_service');
 
         if (1 === intval($step))
         {
             if (!$request->files->has('file'))
                 return new Response("You must upload a file.", 500);
-            $fileCSV = $request->files->get('file');
+            try
+            {
+                $return = $householdService->saveCSV($countryIso3, $project, $request->files->get('file'), $contentJson, $step, $token);
+            }
+            catch (\Exception $e)
+            {
+                dump($e);
+                return new Response($e->getMessage(), Response::HTTP_INTERNAL_SERVER_ERROR);
+            }
         }
         else
         {
-            $fileCSV = null;
+            try
+            {
+                $return = $householdService->foundErrors($countryIso3, $project, $contentJson, $step, $token);
+            }
+            catch (\Exception $e)
+            {
+                dump($e);
+                return new Response($e->getMessage(), Response::HTTP_INTERNAL_SERVER_ERROR);
+            }
         }
-        $contentJson = $request->request->all();
-        $countryIso3 = $contentJson['__country'];
-        unset($contentJson['__country']);
-        /** @var HouseholdCSVService $householeService */
-        $householeService = $this->get('beneficiary.household_csv_service');
-        try
-        {
-            $return = $householeService->saveCSV($countryIso3, $project, $fileCSV, $contentJson, $step);
-        }
-        catch (ValidationException $exception)
-        {
-            return new Response(json_encode(current($exception->getErrors())), Response::HTTP_BAD_REQUEST);
-        }
-        catch (\Exception $e)
-        {
-            return new Response($e->getMessage(), Response::HTTP_INTERNAL_SERVER_ERROR);
-        }
-
+dump($return);
         $json = $this->get('jms_serializer')
-            ->serialize($return, 'json');
+            ->serialize($return, 'json', SerializationContext::create()->setSerializeNull(true)->setGroups(["FullHousehold"]));
         return new Response($json);
     }
