+++ src/BeneficiaryBundle/Entity/Beneficiary.php
@@ -110,7 +110,8 @@ class Beneficiary
      * @Groups({"FullHousehold"})
      */
     private $nationalIds;
-    
+
+
     /**
      * Constructor
      */
