+++ src/BeneficiaryBundle/Resources/config/services.yml
@@ -3,19 +3,31 @@ services:
         class: BeneficiaryBundle\Utils\HouseholdService
         arguments: ["@doctrine.orm.entity_manager", "@jms_serializer", "@beneficiary.beneficiary_service", "@request_validator"]
         public: true
-    beneficiary.household_csv_service:
-        class: BeneficiaryBundle\Utils\HouseholdCSVService
-        arguments: ["@doctrine.orm.entity_manager", "@beneficiary.household_service"]
-        public: true
-    beneficiary.household_export_csv_service:
-        class: BeneficiaryBundle\Utils\ExportCSVService
-        arguments: ["@doctrine.orm.entity_manager", "@service_container"]
-        public: true
     beneficiary.beneficiary_service:
         class: BeneficiaryBundle\Utils\BeneficiaryService
         arguments: ["@doctrine.orm.entity_manager", "@jms_serializer", "@request_validator"]
         public: true
     beneficiary.country_specific_service:
         class: BeneficiaryBundle\Utils\CountrySpecificService
+        arguments: ["@doctrine.orm.entity_manager"]
+        public: true
+
+#####################################################################################################
+####                                            CSV                                              ####
+#####################################################################################################
+    beneficiary.household_export_csv_service:
+        class: BeneficiaryBundle\Utils\ExportCSVService
+        arguments: ["@doctrine.orm.entity_manager", "@service_container"]
+        public: true
+    beneficiary.household_csv_service:
+        class: BeneficiaryBundle\Utils\HouseholdCSVService
+        arguments: ["@doctrine.orm.entity_manager",
+         "@beneficiary.household_service",
+          "@beneficiary.beneficiary_service",
+           "@beneficiary.mapper",
+            "@service_container"]
+        public: true
+    beneficiary.mapper:
+        class: BeneficiaryBundle\Utils\Mapper
         arguments: ["@doctrine.orm.entity_manager"]
         public: true
