+++ src/BeneficiaryBundle/Utils/DataTreatment/InterfaceTreatment.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace BeneficiaryBundle\Utils\DataTreatment;
+
+
+use ProjectBundle\Entity\Project;
+
+interface InterfaceTreatment
+{
+    public function treat(Project $project, array $householdsArray);
+}
