+++ src/BeneficiaryBundle/Utils/DataVerifier/InterfaceVerifier.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace BeneficiaryBundle\Utils\DataVerifier;
+
+
+interface InterfaceVerifier
+{
+    public function verify(string $countryISO3, array $householdArray);
+}
