+++ src/BeneficiaryBundle/Utils/DataVerifier/MoreVerifier.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace BeneficiaryBundle\Utils\DataVerifier;
+
+
+class MoreVerifier extends AbstractVerifier
+{
+
+    public function verify(string $countryISO3, array $householdArray)
+    {
+
+    }
+}
