+++ src/BeneficiaryBundle/Utils/HouseholdCSVService.php
@@ -10,12 +10,20 @@ use BeneficiaryBundle\Entity\Household;
 use BeneficiaryBundle\Entity\VulnerabilityCriterion;
 use BeneficiaryBundle\Model\ImportStatistic;
 use BeneficiaryBundle\Model\IncompleteLine;
+use BeneficiaryBundle\Utils\DataTreatment\AbstractTreatment;
+use BeneficiaryBundle\Utils\DataTreatment\DuplicateTreatment;
+use BeneficiaryBundle\Utils\DataTreatment\TypoTreatment;
+use BeneficiaryBundle\Utils\DataVerifier\AbstractVerifier;
+use BeneficiaryBundle\Utils\DataVerifier\DuplicateVerifier;
+use BeneficiaryBundle\Utils\DataVerifier\TypoVerifier;
 use Doctrine\ORM\EntityManagerInterface;
 use PhpOffice\PhpSpreadsheet\Reader\Csv;
 use PhpOffice\PhpSpreadsheet\Shared\Date;
 use ProjectBundle\Entity\Project;
 use RA\RequestValidatorBundle\RequestValidator\ValidationException;
+use Symfony\Component\DependencyInjection\Container;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\HttpKernel\Kernel;
 
 class HouseholdCSVService
 {
@@ -30,68 +38,32 @@ class HouseholdCSVService
      */
     private $householdService;
 
-    /**
-     * TODO ADJUSTE
-     * Minimum percent to detect a similar household
-     * @var int
-     */
-    private $minimumPercentSimilar = 90;
+    /** @var Mapper $mapper */
+    private $mapper;
 
-    /**
-     * The row index of the header (with the name of country specifics)
-     * @var int
-     */
-    private $indexRowHeader = 2;
+    /** @var Container $container */
+    private $container;
 
-    /**
-     * First row with data
-     * @var int $first_row
-     */
-    private $first_row = 3;
+    /** @var BeneficiaryService $beneficiaryService */
+    private $beneficiaryService;
 
-    /**
-     * First value with a column in the csv which can move, depends on the number of country specifics
-     * @var string
-     */
-    private $firstColumnNonStatic = 'L';
-
-    /**
-     * @var array $MAPPING_CSV
-     */
-    private $MAPPING_CSV = [
-        // Household
-        "address_street" => "A",
-        "address_number" => "B",
-        "address_postcode" => "C",
-        "livelihood" => "D",
-        "notes" => "E",
-        "latitude" => "F",
-        "longitude" => "G",
-        "location" => [
-            // Location
-            "adm1" => "H",
-            "adm2" => "I",
-            "adm3" => "J",
-            "adm4" => "K"
-        ],
-        // Beneficiary
-        "beneficiaries" => [
-            "given_name" => "L",
-            "family_name" => "M",
-            "gender" => "N",
-            "status" => "O",
-            "date_of_birth" => "P",
-            "vulnerability_criteria" => "Q",
-            "phones" => "R",
-            "national_ids" => "S"
-        ]
-    ];
+    /** @var string $token */
+    private $token;
 
 
-    public function __construct(EntityManagerInterface $entityManager, HouseholdService $householdService)
+    public function __construct(
+        EntityManagerInterface $entityManager,
+        HouseholdService $householdService,
+        BeneficiaryService $beneficiaryService,
+        Mapper $mapper,
+        Container $container
+    )
     {
         $this->em = $entityManager;
         $this->householdService = $householdService;
+        $this->beneficiaryService = $beneficiaryService;
+        $this->mapper = $mapper;
+        $this->container = $container;
     }
 
 
@@ -103,12 +75,13 @@ class HouseholdCSVService
      * @param UploadedFile $uploadedFile
      * @param array $contentJson
      * @param int $step
+     * @param $token
      * @return array
      * @throws \Exception
      * @throws \PhpOffice\PhpSpreadsheet\Exception
      * @throws \PhpOffice\PhpSpreadsheet\Reader\Exception
      */
-    public function saveCSV($countryIso3, Project $project, UploadedFile $uploadedFile, array $contentJson, int $step)
+    public function saveCSV($countryIso3, Project $project, UploadedFile $uploadedFile, array $contentJson, int $step, $token)
     {
         // If it's the first step, we transform CSV to array mapped for corresponding to the entity Household
         if ($step === 1)
@@ -120,7 +93,7 @@ class HouseholdCSVService
             $sheetArray = $worksheet->toArray(null, true, true, true);
 
             // Get the list of households from csv with their beneficiaries
-            $listHouseholdsArray = $this->getListHouseholdArray($sheetArray, $countryIso3);
+            $listHouseholdsArray = $this->mapper->getListHouseholdArray($sheetArray, $countryIso3);
         }
         // Else we just get the list of households from the front
         else
@@ -128,63 +101,37 @@ class HouseholdCSVService
             $listHouseholdsArray = $contentJson;
         }
 
-        // Get the list of households from the database with their beneficiaries
-        $listHouseholdsSaved = $this->em->getRepository(Household::class)->getAllBy($countryIso3);
+        return $this->foundErrors($countryIso3, $project, $listHouseholdsArray, $step, $token);
 
-        dump($step);
-        dump($contentJson);
-
-        // TODO DISPATCH 4 STEPS IN 4 CLASS OR NO ?
-
-        switch ($step)
-        {
-            // CASE FOUND TYPO ISSUES
-            case 1:
-                // We update the old with the new, we keep only the old or we add the new
-                break;
-            // CASE FOUND MORE ISSUES
-            case 2:
-                break;
-            // CASE FOUND LESS ISSUES
-            case 3:
-                break;
-            // CASE FOUND DUPLICATED ISSUES
-            case 4:
-                break;
-            // NOT FOUND CASE
-            default:
-                throw new \Exception("Step '$step' unknown.");
-        }
-
-        $errorsArray = $this->foundErrors($countryIso3, $project, $listHouseholdsArray, $listHouseholdsSaved, $step);
-        return $errorsArray;
     }
-
 
     /**
      * @param $countryIso3
      * @param Project $project
      * @param array $listHouseholdsArray
-     * @param array $listHouseholdsSaved
      * @param int $step
+     * @param $token
      * @return array
      * @throws \Exception
      */
-    public function foundErrors($countryIso3, Project $project, array $listHouseholdsArray, array $listHouseholdsSaved, int $step)
+    public function foundErrors($countryIso3, Project $project, array $listHouseholdsArray, int $step, $token)
     {
+        $this->token = $token;
+        dump($listHouseholdsArray);
+        // If there is a treatment class for this step, call it
+        $treatment = $this->guessTraitment($step);
+        if ($treatment !== null)
+            $listHouseholdsArray = $treatment->treat($project, $listHouseholdsArray);
+        dump($listHouseholdsArray);
+
+        $this->getFromCache($step, $listHouseholdsArray);
+        dump($listHouseholdsArray);
+
+        /** @var AbstractVerifier $verifier */
+        $verifier = $this->guessVerifier($step);
+        $return = [];
         $statistic = new ImportStatistic();
-        // List of household which contains the household from the csv and the one which is similar from the database
-        $listHouseholdsTypo = [];
-        // List of household which contains at least one beneficiary who is in another household (in the database)
-        $listHouseholdsDuplicated = [];
-        // List of household which contains more beneficiaries in the csv than is the db
-        $listHouseholdsMoreBeneficiaries = [];
-        // List of household which contains less beneficiaries in the csv than is the db
-        $listHouseholdsLessBeneficiaries = [];
-
-
-        // Variable to specify the line numero where a household is incomplete in CSV
-        $currentLine = $this->first_row;
+        $currentLine = 3;
         foreach ($listHouseholdsArray as $index => $householdArray)
         {
             // If there is a field equal to null, we increment the number of incomplete household and we go to the next household
@@ -195,285 +142,132 @@ class HouseholdCSVService
                 $currentLine += count($householdArray['beneficiaries']);
                 continue;
             }
-            $percent = -1;
-            $similarHousehold = $this->foundSimilarHeadAndHousehold($householdArray, $listHouseholdsSaved, $percent);
-            // We have found a household similar to the current one
-            if ($similarHousehold instanceof Household)
+
+            $returnTmp = $verifier->verify($countryIso3, $householdArray);
+            // IF there is errors
+            if (null != $returnTmp)
             {
-                // Its totally equal
-                if (100 === intval($percent))
-                {
-                    // Check if the beneficiary are all present in both new and old household, with the same name
-                    $similarity = $this->foundSimilarBeneficiaryInHousehold(
-                        $similarHousehold,
-                        $householdArray,
-                        $listHouseholdsMoreBeneficiaries,
-                        $listHouseholdsLessBeneficiaries
-                    );
-
-                    // The household is already in database, so we just add it to the project
-                    if (true === $similarity)
-                    {
-                        $this->householdService->addToProject($similarHousehold, $project);
-                        $statistic->incrementNbAdded();
-                    }
-                    // There is a problem of typo
-                    elseif (false === $similarity)
-                    {
-                        $listHouseholdsTypo[] = ["new" => $householdArray, "old" => $similarHousehold];
-                    }
-
-                    unset($listHouseholdsArray[$index]);
-                }
-                // They are not really the same, return both
+                if ($returnTmp instanceof Household)
+                    $return[] = ["old" => $returnTmp, "new" => $householdArray];
                 else
-                {
-                    $listHouseholdsTypo[] = ["new" => $householdArray, "old" => $similarHousehold];
-                }
+                    $return[] = $returnTmp;
+                unset($listHouseholdsArray[$index]);
             }
-            else
-            {
-                if (($householdDuplicated = $this->checkBeneficiaryInOtherHousehold($householdArray, $countryIso3)) instanceof Household)
-                {
-                    $listHouseholdsDuplicated[] = ["new" => $householdArray, "old" => $householdDuplicated];
-                }
-                else
-                {
-                    try
-                    {
-                        $this->householdService->create($householdArray, $project);
-                        $statistic->incrementNbAdded();
-                    }
-                    catch (\Exception $exception)
-                    {
-                        // If there is a problem during the creation of the household
-                        $statistic->addIncompleteLine(new IncompleteLine($currentLine));
-                    }
-                }
 
-            }
             $currentLine += count($householdArray['beneficiaries']);
         }
 
-        return [
-            "statistic" => $statistic,
-            "typo" => $listHouseholdsTypo,
-            "duplicate" => $listHouseholdsDuplicated,
-            "more" => $listHouseholdsMoreBeneficiaries,
-            "less" => $listHouseholdsLessBeneficiaries
-        ];
+        $this->saveInCache($step, json_encode($listHouseholdsArray));
+
+        return ["data" => $return, "token" => $token];
     }
 
-
     /**
-     * Check if at least one of beneficiary is not assigned to another household
-     *
-     * @param array $newHousehold
-     * @param $countryISO3
-     * @return Household || null
+     * Depends on the step, guess which verifier used
+     * @param int $step
+     * @return AbstractVerifier
+     * @throws \Exception
      */
-    public function checkBeneficiaryInOtherHousehold(array &$newHousehold, $countryISO3)
+    private function guessVerifier(int $step)
     {
-        $oldBeneficiaries = $this->em->getRepository(Beneficiary::class)->findByCriteria($countryISO3, []);
-
-        $newBeneficiaryTmp = null;
-        $oldBeneficiaryTmp = null;
-
-        foreach ($newHousehold['beneficiaries'] as $newBeneficiary)
+        switch ($step)
         {
-            $stringOldHousehold = strtolower(trim($newBeneficiary['given_name']) . "//" . trim($newBeneficiary['family_name']));
-            /** @var Beneficiary $oldBeneficiary */
-            foreach ($oldBeneficiaries as $oldBeneficiary)
-            {
-                if (
-                    strtolower(trim($oldBeneficiary->getGivenName()) . "//" . trim($oldBeneficiary->getFamilyName()))
-                    ===
-                    $stringOldHousehold
-                )
-                {
-                    $newBeneficiaryTmp = $newBeneficiary;
-                    $oldBeneficiaryTmp = $oldBeneficiary;
-                    break;
-                }
-            }
-            if (null !== $newBeneficiaryTmp)
-            {
-                $oldHousehold = $oldBeneficiaryTmp->getHousehold();
-                foreach ($oldHousehold->getBeneficiaries() as $beneficiary)
-                {
-                    if ($beneficiary !== $oldBeneficiaryTmp)
-                        $oldHousehold->removeBeneficiary($beneficiary);
-                }
-
-                foreach ($newHousehold['beneficiaries'] as $index => $beneficiary)
-                {
-                    if ($beneficiary !== $oldBeneficiaryTmp)
-                        unset($newHousehold['beneficiaries'][$index]);
-                }
-
-                return $oldHousehold;
-            }
-        }
-
-        return null;
-    }
-
-
-    /**
-     * If all beneficiaries are equals, return true
-     * If a difference is found in both side, return false
-     * If a difference is found only in one side, add the household to the corresponding array (more beneficiries or less)
-     *
-     * @param Household $oldHousehold
-     * @param array $newHouseholdArray
-     * @param array $listHouseholdsMoreBeneficiaries
-     * @param array $listHouseholdsLessBeneficiaries
-     * @return bool|null
-     */
-    private function foundSimilarBeneficiaryInHousehold(
-        Household $oldHousehold,
-        array $newHouseholdArray,
-        array &$listHouseholdsMoreBeneficiaries,
-        array &$listHouseholdsLessBeneficiaries
-    )
-    {
-        $oldBeneficiaries = $oldHousehold->getBeneficiaries();
-        $newBeneficiaries = $newHouseholdArray['beneficiaries'];
-
-        $oldBeneficiariesAreInNewBeneficiaries = true;
-        $newBeneficiariesAreInOldBeneficiaries = true;
-
-        /** @var Beneficiary $oldBeneficiary */
-        foreach ($oldBeneficiaries as $oldBeneficiary)
-        {
-            $oldBeneficiariesAreInNewBeneficiaries = false;
-            foreach ($newBeneficiaries as $newBeneficiary)
-            {
-                // If the both name are similar, go to the next oldBeneficiary
-                if (
-                    trim($oldBeneficiary->getGivenName()) . '//' . trim($oldBeneficiary->getFamilyName())
-                    ==
-                    trim($newBeneficiary['given_name']) . '//' . trim($newBeneficiary['family_name'])
-                )
-                {
-                    $oldBeneficiariesAreInNewBeneficiaries = true;
-                    break;
-                }
-            }
-            if (!$oldBeneficiariesAreInNewBeneficiaries)
-            {
+            // CASE FOUND TYPO ISSUES
+            case 1:
+                return new TypoVerifier($this->em);
                 break;
-            }
-        }
-        foreach ($newBeneficiaries as $newBeneficiary)
-        {
-            $newBeneficiariesAreInOldBeneficiaries = false;
-            foreach ($oldBeneficiaries as $oldBeneficiary)
-            {
-                // If the both name are similar, go to the next $newBeneficiary
-                if (
-                    trim($oldBeneficiary->getGivenName()) . '//' . trim($oldBeneficiary->getFamilyName())
-                    ==
-                    trim($newBeneficiary['given_name']) . '//' . trim($newBeneficiary['family_name'])
-                )
-                {
-                    $newBeneficiariesAreInOldBeneficiaries = true;
-                    break;
-                }
-            }
-            if (!$newBeneficiariesAreInOldBeneficiaries)
-            {
+            // CASE FOUND MORE ISSUES
+            case 2:
+                return new DuplicateVerifier($this->em);
                 break;
-            }
-        }
-
-        if ($oldBeneficiariesAreInNewBeneficiaries && $newBeneficiariesAreInOldBeneficiaries)
-        {
-            return true;
-        }
-        elseif ($oldBeneficiariesAreInNewBeneficiaries)
-        {
-            $listHouseholdsMoreBeneficiaries[] = ["old" => $oldHousehold, "new" => $newHouseholdArray];
-            return null;
-        }
-        elseif ($newBeneficiariesAreInOldBeneficiaries)
-        {
-            $listHouseholdsLessBeneficiaries[] = ["old" => $oldHousehold, "new" => $newHouseholdArray];
-            return null;
-        }
-        else
-        {
-            return false;
+            // CASE FOUND LESS ISSUES
+            case 3:
+                throw new \Exception("To be implemented");
+                break;
+            // CASE FOUND DUPLICATED ISSUES
+            case 4:
+                throw new \Exception("To be implemented");
+                break;
+            // NOT FOUND CASE
+            default:
+                throw new \Exception("Step '$step' unknown.");
         }
     }
 
     /**
-     * Found if a household and its beneficiaries are similar to these from $newHouseholdArray
-     * @param array $newHouseholdArray
-     * @param array $listHouseholds
-     * @param int $percent
-     * @return array
+     * Depends on the step, guess which verifier used
+     * @param int $step
+     * @return AbstractTreatment
+     * @throws \Exception
      */
-    private function foundSimilarHeadAndHousehold(array $newHouseholdArray, array $listHouseholds, int &$percent)
+    private function guessTraitment(int $step)
     {
-        $newHead = null;
-        foreach ($newHouseholdArray['beneficiaries'] as $newBeneficiaryArray)
+        switch ($step)
         {
-            if (1 === intval($newBeneficiaryArray['status']))
-            {
-                $newHead = $newBeneficiaryArray;
+            // CASE FOUND TYPO ISSUES
+            case 1:
+                return null;
                 break;
-            }
+            // CASE FOUND MORE ISSUES
+            case 2:
+                return new TypoTreatment($this->em, $this->householdService, $this->beneficiaryService, $this->container, $this->token);
+                break;
+            // CASE FOUND LESS ISSUES
+            case 3:
+                return new DuplicateTreatment($this->em, $this->householdService, $this->beneficiaryService);
+                break;
+            // CASE FOUND DUPLICATED ISSUES
+            case 4:
+                throw new \Exception("To be implemented");
+                break;
+            // NOT FOUND CASE
+            default:
+                throw new \Exception("Step '$step' unknown.");
         }
-        if (null === $newHead)
-            return null;
-
-        // Concatenation of fields to compare with
-        $stringNewHouseholdToCompare = $newHouseholdArray["address_street"] . "//" .
-            $newHouseholdArray["address_number"] . "//" .
-            $newHouseholdArray["address_postcode"] . "//" .
-            $newHead["given_name"] . "//" .
-            $newHead["family_name"];
-
-        $similarHousehold = null;
-        $percent = $this->minimumPercentSimilar;
-        /** @var Household $household */
-        foreach ($listHouseholds as $oldHousehold)
-        {
-            // Get the head of the current household
-            /** @var Beneficiary $oldHead */
-            $oldHead = $this->em->getRepository(Beneficiary::class)->getHeadOfHousehold($oldHousehold);
-            if (!$oldHead instanceof Beneficiary)
-                continue;
-
-            $stringOldHouseholdToCompare = $oldHousehold->getAddressStreet() . "//" .
-                $oldHousehold->getAddressNumber() . "//" .
-                $oldHousehold->getAddressPostcode() . "//" .
-                $oldHead->getGivenName() . "//" .
-                $oldHead->getFamilyName();
-
-
-            similar_text(
-                $stringNewHouseholdToCompare,
-                $stringOldHouseholdToCompare,
-                $tmpPercent
-            );
-
-            if (100 == $tmpPercent)
-            {
-                $percent = $tmpPercent;
-                return $oldHousehold;
-            }
-            elseif ($percent < $tmpPercent)
-            {
-                $similarHousehold = $oldHousehold;
-                $percent = $tmpPercent;
-            }
-        }
-
-        return $similarHousehold;
     }
 
+    /**
+     * @param int $step
+     * @param array $listHouseholdsArray
+     * @param string $token
+     * @throws \Exception
+     */
+    private function getFromCache(int $step, array &$listHouseholdsArray)
+    {
+        if ($step <= 1 || null === $this->token)
+            return;
+
+        $dir_root = $this->container->get('kernel')->getRootDir();
+        $dir_var = $dir_root . '/../var/data/' . $this->token;
+        if (!is_dir($dir_var))
+            mkdir($dir_var);
+
+        $fileContent = file_get_contents($dir_var . '/step_' . strval($step - 1));
+        $householdsCached = json_decode($fileContent, true);
+        foreach ($householdsCached as $householdCached)
+        {
+            $listHouseholdsArray[] = $householdCached;
+        }
+    }
+
+    /**
+     * @param int $step
+     * @param $dataToSave
+     * @param string|null $token
+     * @throws \Exception
+     */
+    private function saveInCache(int $step, $dataToSave)
+    {
+        $sizeToken = 50;
+        if (null === $this->token)
+            $this->token = bin2hex(random_bytes($sizeToken));
+
+        $dir_root = $this->container->get('kernel')->getRootDir();
+        $dir_var = $dir_root . '/../var/data/' . $this->token;
+        if (!is_dir($dir_var))
+            mkdir($dir_var);
+        file_put_contents($dir_var . '/step_' . $step, $dataToSave);
+    }
 
     /**
      * Check if a value is missing inside the array
@@ -496,282 +290,4 @@ class HouseholdCSVService
 
         return $isIncomplete;
     }
-
-    /**
-     * Get the list of households with their beneficiaries
-     * @param array $sheetArray
-     * @param $countryIso3
-     * @return array
-     */
-    private function getListHouseholdArray(array $sheetArray, $countryIso3)
-    {
-        // Get the mapping for the current country
-        $mappingCSV = $this->loadMappingCSVOfCountry($countryIso3);
-        $listHouseholdArray = [];
-        $householdArray = null;
-        $rowHeader = null;
-        $formattedHouseholdArray = null;
-
-        foreach ($sheetArray as $indexRow => $row)
-        {
-            if ($this->indexRowHeader === $indexRow)
-                $rowHeader = $row;
-            if ($indexRow < $this->first_row)
-                continue;
-
-            // Load the household array for the current row
-            $formattedHouseholdArray = $this->mappingCSV($mappingCSV, $countryIso3, $row, $rowHeader);
-            // Check if it's a new household or just a new beneficiary in the current row
-            if ($formattedHouseholdArray["address_street"] !== null)
-            {
-                if (null !== $householdArray)
-                {
-                    $listHouseholdArray[] = $householdArray;
-                }
-                $householdArray = $formattedHouseholdArray;
-            }
-            else
-            {
-                $householdArray["beneficiaries"][] = current($formattedHouseholdArray["beneficiaries"]);
-            }
-        }
-
-        if (null !== $formattedHouseholdArray)
-        {
-            $listHouseholdArray[] = $householdArray;
-        }
-
-        return $listHouseholdArray;
-    }
-
-    /**
-     * Transform the array from the CSV (with index 'A', 'B') to a formatted array which can be compatible with the
-     * function save of a household (with correct index names and correct deep array)
-     *
-     * @param array $mappingCSV
-     * @param $countryIso3
-     * @param array $row
-     * @param array $rowHeader
-     * @return array
-     */
-    private function mappingCSV(array $mappingCSV, $countryIso3, array $row, array $rowHeader)
-    {
-        $formattedHouseholdArray = [];
-        foreach ($mappingCSV as $formattedIndex => $csvIndex)
-        {
-            if (is_array($csvIndex))
-            {
-                foreach ($csvIndex as $formattedIndex2 => $csvIndex2)
-                {
-                    if (null !== $row[$csvIndex2])
-                        $row[$csvIndex2] = strval($row[$csvIndex2]);
-                    $formattedHouseholdArray[$formattedIndex][$formattedIndex2] = $row[$csvIndex2];
-                }
-            }
-            else
-            {
-                if (null !== $row[$csvIndex])
-                    $row[$csvIndex] = strval($row[$csvIndex]);
-                $formattedHouseholdArray[$formattedIndex] = $row[$csvIndex];
-            }
-        }
-        // Add the country iso3 from the request
-        $formattedHouseholdArray["location"]["country_iso3"] = $countryIso3;
-
-        // Traitment on field with multiple value or foreign key inside (switch name to id for example)
-        $this->fieldCountrySpecifics($mappingCSV, $formattedHouseholdArray, $rowHeader);
-        $this->fieldVulnerabilityCriteria($formattedHouseholdArray);
-        $this->fieldPhones($formattedHouseholdArray);
-        $this->fieldNationalIds($formattedHouseholdArray);
-        $this->fieldBeneficiary($formattedHouseholdArray);
-
-        // ADD THE FIELD COUNTRY ONLY FOR THE CHECKING BY THE REQUEST VALIDATOR
-        $formattedHouseholdArray["__country"] = $countryIso3;
-        return $formattedHouseholdArray;
-    }
-
-    /**
-     * Reformat the fields countries_specific_answers
-     * @param array $mappingCSV
-     * @param $formattedHouseholdArray
-     * @param array $rowHeader
-     */
-    private function fieldCountrySpecifics(array $mappingCSV, &$formattedHouseholdArray, array $rowHeader)
-    {
-        $formattedHouseholdArray["country_specific_answers"] = [];
-        foreach ($formattedHouseholdArray as $indexFormatted => $value)
-        {
-            if (substr($indexFormatted, 0, 20) === "tmp_country_specific")
-            {
-                $field = $rowHeader[$mappingCSV[$indexFormatted]];
-                $countrySpecific = $this->em->getRepository(CountrySpecific::class)
-                    ->findOneByField($field);
-                $formattedHouseholdArray["country_specific_answers"][] = [
-                    "answer" => $value,
-                    "country_specific" => ["id" => $countrySpecific->getId()]
-                ];
-                unset($formattedHouseholdArray[$indexFormatted]);
-            }
-        }
-    }
-
-    /**
-     * Reformat the field which contains vulnerability criteria => switch list of names to a list of ids
-     * @param $formattedHouseholdArray
-     */
-    private function fieldVulnerabilityCriteria(&$formattedHouseholdArray)
-    {
-        $vulnerability_criteria_string = $formattedHouseholdArray["beneficiaries"]["vulnerability_criteria"];
-        $vulnerability_criteria_array = array_map('trim', explode(";", $vulnerability_criteria_string));
-        $formattedHouseholdArray["beneficiaries"]["vulnerability_criteria"] = [];
-        foreach ($vulnerability_criteria_array as $item)
-        {
-            $vulnerability_criterion = $this->em->getRepository(VulnerabilityCriterion::class)->findOneByValue($item);
-            if (!$vulnerability_criterion instanceof VulnerabilityCriterion)
-                continue;
-            $formattedHouseholdArray["beneficiaries"]["vulnerability_criteria"][] = ["id" => $vulnerability_criterion->getId()];
-        }
-    }
-
-    /**
-     * Reformat the field phones => switch string 'type-number' to [type => type, number => number]
-     * @param $formattedHouseholdArray
-     */
-    private function fieldPhones(&$formattedHouseholdArray)
-    {
-        $phones_string = $formattedHouseholdArray["beneficiaries"]["phones"];
-        $phones_array = array_map('trim', explode(";", $phones_string));
-        $formattedHouseholdArray["beneficiaries"]["phones"] = [];
-        foreach ($phones_array as $item)
-        {
-            if ("" == $item)
-                continue;
-            $item_array = array_map('trim', explode("-", $item));
-            $formattedHouseholdArray["beneficiaries"]["phones"][] = ["type" => $item_array[0], "number" => $item_array[1]];
-        }
-    }
-
-    /**
-     * Reformat the field nationalids => switch string 'idtype-idnumber' to [id_type => idtype, id_number => idnumber]
-     * @param $formattedHouseholdArray
-     */
-    private function fieldNationalIds(&$formattedHouseholdArray)
-    {
-        $national_ids_string = $formattedHouseholdArray["beneficiaries"]["national_ids"];
-        $national_ids_array = array_map('trim', explode(";", $national_ids_string));
-        $formattedHouseholdArray["beneficiaries"]["national_ids"] = [];
-        foreach ($national_ids_array as $item)
-        {
-            if ("" == $item)
-                continue;
-            $item_array = array_map('trim', explode("-", $item));
-            $formattedHouseholdArray["beneficiaries"]["national_ids"][] = ["id_type" => $item_array[0], "id_number" => $item_array[1]];
-        }
-    }
-
-    /**
-     * Load the mapping CSV for a specific country. Some columns can move because on the number of country specifics
-     *
-     * @param $countryIso3
-     * @return array
-     */
-    private function loadMappingCSVOfCountry($countryIso3)
-    {
-        $countrySpecifics = $this->em->getRepository(CountrySpecific::class)->findByCountryIso3($countryIso3);
-        // Get the number of country specific for the specific country countryIso3
-        $nbCountrySpecific = count($countrySpecifics);
-        $mappingCSVCountry = [];
-        $countrySpecificsAreLoaded = false;
-        foreach ($this->MAPPING_CSV as $indexFormatted => $indexCSV)
-        {
-            // For recursive array (allowed only 1 level of recursivity)
-            if (is_array($indexCSV))
-            {
-                foreach ($indexCSV as $indexFormatted2 => $indexCSV2)
-                {
-                    // If the column is before the non-static columns, change nothing
-                    if ($indexCSV2 < $this->firstColumnNonStatic)
-                        $mappingCSVCountry[$indexFormatted][$indexFormatted2] = $indexCSV2;
-                    // Else we increment the column.
-                    // Example : if $nbCountrySpecific = 1, we shift the column by 1 (if the column is X, it will became Y)
-                    else
-                    {
-                        // If we have not added the country specific column in the mapping
-                        if (!$countrySpecificsAreLoaded)
-                        {
-                            // Add each country specific column in the mapping
-                            for ($i = 0; $i < $nbCountrySpecific; $i++)
-                            {
-                                $mappingCSVCountry["tmp_country_specific" . $i] =
-                                    $this->SUMOfLetter($indexCSV2, $i);
-                            }
-                            $countrySpecificsAreLoaded = true;
-                        }
-                        $mappingCSVCountry[$indexFormatted][$indexFormatted2] = $this->SUMOfLetter($indexCSV2, $nbCountrySpecific);
-                    }
-                }
-            }
-            else
-            {
-                // Same process than in the if
-                if ($indexCSV < $this->firstColumnNonStatic)
-                    $mappingCSVCountry[$indexFormatted] = $indexCSV;
-                else
-                {
-                    // If we have not added the country specific column in the mapping
-                    if (!$countrySpecificsAreLoaded)
-                    {
-                        // Add each country specific column in the mapping
-                        for ($i = 0; $i < $nbCountrySpecific; $i++)
-                        {
-                            $mappingCSVCountry["tmp_country_specific" . $i] =
-                                $this->SUMOfLetter($indexCSV, $i);
-                        }
-                        $countrySpecificsAreLoaded = true;
-                    }
-                    $mappingCSVCountry[$indexFormatted] = $this->SUMOfLetter($indexCSV, $nbCountrySpecific);
-                }
-            }
-        }
-
-        return $mappingCSVCountry;
-    }
-
-    /**
-     * Reformat the field beneficiary
-     * @param $formattedHouseholdArray
-     */
-    private function fieldBeneficiary(&$formattedHouseholdArray)
-    {
-        $beneficiary = $formattedHouseholdArray["beneficiaries"];
-        $beneficiary["profile"] = ["photo" => ""];
-        $beneficiary["updated_on"] = (new \DateTime())->format('Y-m-d H:m:i');
-        unset($formattedHouseholdArray["beneficiaries"]);
-        $formattedHouseholdArray["beneficiaries"][] = $beneficiary;
-    }
-
-    /**
-     * Make an addition of a letter and a number
-     * Example : A + 2 = C  Or  Z + 1 = AA  OR  AY + 2 = BA
-     * @param $letter1
-     * @param $number
-     * @return string
-     */
-    private function SUMOfLetter($letter1, $number)
-    {
-        $ascii = ord($letter1) + $number;
-        $prefix = '';
-        if ($ascii > 90)
-        {
-            $prefix = 'A';
-            $ascii -= 26;
-            while ($ascii > 90)
-            {
-                $prefix++;
-                $ascii -= 90;
-            }
-        }
-        return $prefix . chr($ascii);
-    }
-
 }
